Distill Outcomes

distill takes a low-age resolution outcome, for example deaths, and proportionally distributes that outcome into a higher age resolution for use in subsequent analyses like years-life-lost style calculations.

Usage

distill(alembic_dt, outcomes_dt, groupcol = names(outcomes_dt)[1])

Arguments

alembic_dt

an alembic() return value

outcomes_dt

a long-format data.frame with a column either named from or model_from and a column value (other columns will be silently ignored)

groupcol

a string, the name of the outcome model group column. The outcomes_dt[[groupcol]] column must match the model_partition lower bounds, as provided when constructing the alembic_dt with alembic().

Value

a data.frame, with output_partition and recalculated value column

Details

When the value column is re-calculated, note that it will aggregate all rows with matching groupcol entries in outcomes_dt. If you need to group by other features in your input data (e.g. if you need to distill outcomes across multiple simulation outputs or at multiple time points), that has to be done by external grouping then calling distill().

Examples

ifr_levin <- function(age_in_years) {
  (10^(-3.27 + 0.0524 * age_in_years))/100
}
age_limits <- c(seq(0, 69, by = 5), 70, 80, 100)
age_pyramid <- data.table(
  from = 0:99, weight = ifelse(0:99 < 65, 1, .99^(0:99-64))
) # flat age distribution, then 1% annual deaths
alembic_dt <- alembic(ifr_levin, age_pyramid, age_limits, 0:100)
#> Warning: data may not support bounds: upper bound 0 > last column 1 entry 0

results <- data.table(model_partition = head(age_limits, -1))[, value := 10]
distill(alembic_dt, results)
#>      output_partition     value
#>                 <int>     <num>
#>   1:                0 1.5485580
#>   2:                1 1.7471391
#>   3:                2 1.9711854
#>   4:                3 2.2239626
#>   5:                4 2.5091549
#>   6:                5 1.5485580
#>   7:                6 1.7471391
#>   8:                7 1.9711854
#>   9:                8 2.2239626
#>  10:                9 2.5091549
#>  11:               10 1.5485580
#>  12:               11 1.7471391
#>  13:               12 1.9711854
#>  14:               13 2.2239626
#>  15:               14 2.5091549
#>  16:               15 1.5485580
#>  17:               16 1.7471391
#>  18:               17 1.9711854
#>  19:               18 2.2239626
#>  20:               19 2.5091549
#>  21:               20 1.5485580
#>  22:               21 1.7471391
#>  23:               22 1.9711854
#>  24:               23 2.2239626
#>  25:               24 2.5091549
#>  26:               25 1.5485580
#>  27:               26 1.7471391
#>  28:               27 1.9711854
#>  29:               28 2.2239626
#>  30:               29 2.5091549
#>  31:               30 1.5485580
#>  32:               31 1.7471391
#>  33:               32 1.9711854
#>  34:               33 2.2239626
#>  35:               34 2.5091549
#>  36:               35 1.5485580
#>  37:               36 1.7471391
#>  38:               37 1.9711854
#>  39:               38 2.2239626
#>  40:               39 2.5091549
#>  41:               40 1.5485580
#>  42:               41 1.7471391
#>  43:               42 1.9711854
#>  44:               43 2.2239626
#>  45:               44 2.5091549
#>  46:               45 1.5485580
#>  47:               46 1.7471391
#>  48:               47 1.9711854
#>  49:               48 2.2239626
#>  50:               49 2.5091549
#>  51:               50 1.5485580
#>  52:               51 1.7471391
#>  53:               52 1.9711854
#>  54:               53 2.2239626
#>  55:               54 2.5091549
#>  56:               55 1.5485580
#>  57:               56 1.7471391
#>  58:               57 1.9711854
#>  59:               58 2.2239626
#>  60:               59 2.5091549
#>  61:               60 1.5485580
#>  62:               61 1.7471391
#>  63:               62 1.9711854
#>  64:               63 2.2239626
#>  65:               64 2.5091549
#>  66:               65 1.5836554
#>  67:               66 1.7688698
#>  68:               67 1.9757458
#>  69:               68 2.2068167
#>  70:               69 2.4649122
#>  71:               70 0.5782919
#>  72:               71 0.6459253
#>  73:               72 0.7214687
#>  74:               73 0.8058472
#>  75:               74 0.9000941
#>  76:               75 1.0053635
#>  77:               76 1.1229445
#>  78:               77 1.2542771
#>  79:               78 1.4009695
#>  80:               79 1.5648182
#>  81:               80 0.1629244
#>  82:               81 0.1819790
#>  83:               82 0.2032622
#>  84:               83 0.2270344
#>  85:               84 0.2535870
#>  86:               85 0.2832449
#>  87:               86 0.3163715
#>  88:               87 0.3533723
#>  89:               88 0.3947005
#>  90:               89 0.4408622
#>  91:               90 0.4924227
#>  92:               91 0.5500134
#>  93:               92 0.6143395
#>  94:               93 0.6861889
#>  95:               94 0.7664412
#>  96:               95 0.8560794
#>  97:               96 0.9562011
#>  98:               97 1.0680324
#>  99:               98 1.1929429
#> 100:               99 0.0000000
#>      output_partition     value